Fix problem with incorrectly written MMPLD file format version.
git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/trunk@5142 a63bd714-7e14-4b5e-94e7-302c8c8ff188

+++ src/io/MmpldWriter.h
@@ -120,7 +120,7 @@ protected:
 	bool _bWriteControlPrepared;
 
 	long _fileCount;
-	int _mmpldversion;
+	uint16_t _mmpldversion;
 	MMPLD_Vertex_type _vertex_type;
 	MMPLD_Color_type _color_type;
